pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "banking-rm-agent"
version = "1.0.0"
description = "AI-powered Banking Relationship Manager Agent built with CrewAI"
readme = "docs/README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Banking RM Agent Team", email = "support@bankingrmagent.com"},
]
keywords = ["banking", "ai", "crewai", "relationship-manager", "chatbot", "fintech"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "crewai>=0.28.8",
    "crewai-tools>=0.1.6",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "python-dotenv>=1.0.0",
    "streamlit>=1.29.0",
    "pandas>=2.1.4",
    "requests>=2.31.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
]

[project.scripts]
banking-rm-agent = "banking_rm_agent.interfaces.streamlit_app:main"
banking-rm-api = "banking_rm_agent.interfaces.api_server:app"

[project.urls]
Homepage = "https://github.com/your-org/banking-rm-agent"
Documentation = "https://banking-rm-agent.readthedocs.io/"
Repository = "https://github.com/your-org/banking-rm-agent.git"
"Bug Tracker" = "https://github.com/your-org/banking-rm-agent/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
banking_rm_agent = ["config/*.yaml", "config/*.json"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=src/banking_rm_agent --cov-report=html --cov-report=term-missing"
